requirements.txt
selenium
pyautogui
python-dotenv
